Use docker.yaml to configure SkyPortal Dockerfile
This is the way the SP Dockerfile was intended to be configured.
Inside, it copies this to config.yaml and then fills out templates and
sets up services.

If `config.yaml` is used instead, like we did in Fritz, it gets
overwritten by `docker.yaml` at some point—no good.

Note that now:
- fritz build produces docker.yaml
- this file gets copied into the Docker image during build and is
  used to fill out templates / set up services
- is additionally mounted onto /skyportal/config.yaml by docker-compose,
  so that you can make modifications to the config and still test the
  system locally